pyproject.toml
[project]
name = "bookstack-editor"
requires-python = ">=3.10"
version = "0.2.0"
description = ""
authors = [{name = "Robert Krahn", email = "robert@kra.hn"}]
dependencies = [
  "pydantic",
  "loguru",
  "bookstack",
  "tqdm",
  "python-dotenv",
]

[project.scripts]
bookstack-editor = "bookstack_editor.__main__:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["bookstack_editor"]

[tool.pyright]
exclude = []
